Refactor postgres store and remove sqlite version
