+++ lib/entity_store/postgres_entity_store.rb
@@ -10,15 +10,38 @@ module EntityStore
     Sequel.extension :pg_json_ops
 
     class << self
-      attr_accessor :connection_profile
+      attr_accessor :connection_string
       attr_writer :connect_timeout
 
       def database
-        @_database ||= Sequel.connect('postgres://localhost/cronofy_test')
+        @_database ||= Sequel.connect(connection_string)
         @_database.extension :pg_array
         @_database.extension :pg_json
 
         @_database
+      end
+
+      def init
+        db = database
+        unless db.table_exists?(:entities)
+          db.create_table :entities do
+            String :id
+            String :_type
+            integer :snapshot_key
+            integer :version
+            column :snapshot, :jsonb
+          end
+        end
+
+        unless db.table_exists?(:entity_events)
+          db.create_table :entity_events do
+            String :id
+            String :_type
+            String :_entity_id
+            integer :entity_version
+            column :data, :jsonb
+          end
+        end
       end
     end
 
@@ -27,34 +50,10 @@ module EntityStore
     end
 
     def entities
-      return @entities_collection if @entities_collection
-
-      unless open.table_exists?(:entities)
-        open.create_table :entities do
-          String :id
-          String :_type
-          integer :snapshot_key
-          integer :version
-          column :snapshot, :jsonb
-        end
-      end
-
       @entities_collection ||= open[:entities]
     end
 
     def events
-      return @events_collection if @events_collection
-
-      unless open.table_exists?(:entity_events)
-        open.create_table :entity_events do
-          String :id
-          String :_type
-          String :_entity_id
-          integer :entity_version
-          column :data, :jsonb
-        end
-      end
-
       @events_collection ||= open[:entity_events]
     end
 
@@ -65,8 +64,6 @@ module EntityStore
     end
 
     def ensure_indexes
-      #events.ensure_index([['_entity_id', Mongo::ASCENDING], ['_id', Mongo::ASCENDING]])
-      #events.ensure_index([['_entity_id', Mongo::ASCENDING], ['entity_version', Mongo::ASCENDING], ['_id', Mongo::ASCENDING]])
     end
 
     def add_entity(entity)
@@ -118,8 +115,6 @@ module EntityStore
     end
 
     def add_events(items)
-      # Hack to ensure that pg_json is loaded :(
-      open
       items.each do |event|
         doc = {
           :id => BSON::ObjectId.new.to_s,
@@ -132,167 +127,6 @@ module EntityStore
       end
     end
 
-    class TypedJSON
-      def self.generate(obj, *args)
-        hash_dup = each_with_parent(obj)
-        JSON.generate(hash_dup, *args)
-      end
-
-      def self.map_to_json(obj)
-        case obj
-        when Time
-          JSONTime.new(obj)
-        when Date
-          JSONDate.new(obj)
-        when DateTime
-          JSONDateTime.new(obj)
-        when Symbol
-          JSONSymbol.new(obj)
-        else
-          obj
-        end
-      end
-
-      def self.each_with_parent(hash, result=nil)
-        duplicated_hash = {} || result
-
-        hash.each do |k, v|
-          case v
-          when Hash
-            duplicated_hash[k] = each_with_parent(v, duplicated_hash)
-          else
-            duplicated_hash[k] = map_to_json(v)
-          end
-        end
-
-        duplicated_hash
-      end
-    end
-
-    class JSONSymbol < SimpleDelegator
-      # Returns a hash, that will be turned into a JSON object and represent this
-      # object.
-      def as_json(*)
-        {
-          JSON.create_id => self.class.name,
-          's'            => to_s,
-        }
-      end
-
-      # Stores class name (Symbol) with String representation of Symbol as a JSON string.
-      def to_json(*a)
-        as_json.to_json(*a)
-      end
-
-      # Deserializes JSON string by converting the <tt>string</tt> value stored in the object to a Symbol
-      def self.json_create(o)
-        o['s'].to_sym
-      end
-    end
-
-    class JSONTime < SimpleDelegator
-      # Deserializes JSON string by converting time since epoch to Time
-      def self.json_create(object)
-        if usec = object.delete('u') # used to be tv_usec -> tv_nsec
-          object['n'] = usec * 1000
-        end
-        if method_defined?(:tv_nsec)
-          Time.at(object['s'], Time.Rational(object['n'], 1000))
-        else
-          Time.at(object['s'], object['n'] / 1000)
-        end
-      end
-
-      # Returns a hash, that will be turned into a JSON object and represent this
-      # object.
-      def as_json(*)
-        nanoseconds = [ tv_usec * 1000 ]
-        respond_to?(:tv_nsec) and nanoseconds << tv_nsec
-        nanoseconds = nanoseconds.max
-        {
-          JSON.create_id => self.class.name,
-          's'            => tv_sec,
-          'n'            => nanoseconds,
-        }
-      end
-
-      # Stores class name (Time) with number of seconds since epoch and number of
-      # microseconds for Time as JSON string
-      def to_json(*args)
-        as_json.to_json(*args)
-      end
-    end
-
-    class JSONDate < SimpleDelegator
-      # Deserializes JSON string by converting Julian year <tt>y</tt>, month
-      # <tt>m</tt>, day <tt>d</tt> and Day of Calendar Reform <tt>sg</tt> to Date.
-      def self.json_create(object)
-        Date.civil(*object.values_at('y', 'm', 'd', 'sg'))
-      end
-
-      #alias start sg unless method_defined?(:start)
-
-      # Returns a hash, that will be turned into a JSON object and represent this
-      # object.
-      def as_json(*)
-        {
-          JSON.create_id => self.class.name,
-          'y' => year,
-          'm' => month,
-          'd' => day,
-          'sg' => start,
-        }
-      end
-
-      # Stores class name (Date) with Julian year <tt>y</tt>, month <tt>m</tt>, day
-      # <tt>d</tt> and Day of Calendar Reform <tt>sg</tt> as JSON string
-      def to_json(*args)
-        as_json.to_json(*args)
-      end
-    end
-
-    class JSONDateTime < SimpleDelegator
-      # Deserializes JSON string by converting year <tt>y</tt>, month <tt>m</tt>,
-      # day <tt>d</tt>, hour <tt>H</tt>, minute <tt>M</tt>, second <tt>S</tt>,
-      # offset <tt>of</tt> and Day of Calendar Reform <tt>sg</tt> to DateTime.
-      def self.json_create(object)
-        args = object.values_at('y', 'm', 'd', 'H', 'M', 'S')
-        of_a, of_b = object['of'].split('/')
-        if of_b and of_b != '0'
-          args << DateTime.Rational(of_a.to_i, of_b.to_i)
-        else
-          args << of_a
-        end
-        args << object['sg']
-        DateTime.civil(*args)
-      end
-
-      #alias start sg unless method_defined?(:start)
-
-      # Returns a hash, that will be turned into a JSON object and represent this
-      # object.
-      def as_json(*)
-        {
-          JSON.create_id => self.class.name,
-          'y' => year,
-          'm' => month,
-          'd' => day,
-          'H' => hour,
-          'M' => min,
-          'S' => sec,
-          'of' => offset.to_s,
-          'sg' => start,
-        }
-      end
-
-      # Stores class name (DateTime) with Julian year <tt>y</tt>, month <tt>m</tt>,
-      # day <tt>d</tt>, hour <tt>H</tt>, minute <tt>M</tt>, second <tt>S</tt>,
-      # offset <tt>of</tt> and Day of Calendar Reform <tt>sg</tt> as JSON string
-      def to_json(*args)
-        as_json.to_json(*args)
-      end
-    end
-
     # Public: loads the entity from the store, including any available snapshots
     # then loads the events to complete the state
     #
@@ -302,8 +136,7 @@ module EntityStore
     #
     # Returns an array of entities
     def get_entities(ids, options={})
-      # not actually needed
-      object_ids = ids.map do |id|
+      ids.each do |id|
         begin
           BSON::ObjectId.from_string(id)
         rescue BSON::InvalidObjectId
@@ -357,38 +190,26 @@ module EntityStore
     def get_events(criteria)
       return {} if criteria.empty?
 
-      query_items = criteria.map do |item|
+      result = {}
+
+      criteria.each do |item|
         raise ArgumentError.new(":id missing from criteria") unless item[:id]
-        unless [:since_version]
-          events.where(:_entity_id => item[:id]).to_a
-        else
-          events.where(:_entity_id => item[:id]).where('entity_version > ?', item[:since_version]).to_a
+
+        query = events.where(:_entity_id => item[:id])
+
+        if item[:since_version]
+          query = query.where('entity_version > ?', item[:since_version])
         end
-      end.flatten(1)
 
-      result = Hash[ criteria.map { |item| [ item[:id], [] ] } ]
-
-      query_items.each do |attrs|
-        result[attrs[:_entity_id].to_s] << attrs
-      end
-
-      result.each do |_, events|
-        # Have to do the sort client side as otherwise the query will not use
-        # indexes (http://docs.mongodb.org/manual/reference/operator/query/or/#or-and-sort-operations)
-        events.sort_by! { |attrs| [attrs[:entity_version], attrs[:_id].to_s] }
-
-        # Convert the attributes into event objects
-        events.map! do |attrs|
+        result[item[:id]] = query.order(:entity_version, :id).map do |attrs|
           begin
             hash = JSON.load(Sequel.object_to_json(attrs[:data]))
             EntityStore::Config.load_type(attrs[:_type]).new(hash)
           rescue => e
             log_error "Error loading type #{attrs[:_type]}", e
-            nil
+            next
           end
         end
-
-        events.compact!
       end
 
       result
